requirements.txt
altgraph==0.17.4
future==1.0.0
iso8601==2.1.0
packaging==25.0
pefile==2023.2.7
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyserial==3.5
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==80.9.0
shiboken6==6.9.1
